Added Data Validation functionality

+++ src/main/java/com/qa/ims/controller/CustomerController.java
@@ -47,7 +47,6 @@ public class CustomerController implements CrudController<Customer> {
 		String firstName = utils.getString();
 		LOGGER.info("Please enter a surname: ");
 		String surname = utils.getString();
-		// enter address once everything is working
 		Customer customer = customerDAO.create(new Customer(firstName, surname));
 		LOGGER.info("Customer created.");
 		return customer;
@@ -60,6 +59,9 @@ public class CustomerController implements CrudController<Customer> {
 	public Customer update() {
 		LOGGER.info("Please enter the id of the customer you would like to update: ");
 		Long id = utils.getLong();
+
+		if (!checkID(id)) return null;
+
 		LOGGER.info("Please enter a first name: ");
 		String firstName = utils.getString();
 		LOGGER.info("Please enter a surname: ");
@@ -67,6 +69,7 @@ public class CustomerController implements CrudController<Customer> {
 		Customer customer = customerDAO.update(new Customer(id, firstName, surname));
 		LOGGER.info("Customer successfully updated.");
 		return customer;
+
 	}
 
 	/**
@@ -78,7 +81,20 @@ public class CustomerController implements CrudController<Customer> {
 	public int delete() {
 		LOGGER.info("Please enter the ID of the customer you would like to delete: ");
 		Long id = utils.getLong();
+
+		if (!checkID(id)) return 0;
+
+		LOGGER.info("Customer successfully deleted");
 		return customerDAO.delete(id);
+	}
+
+	private boolean checkID(Long id) {
+		List<Customer> customers = customerDAO.readAll();
+		if (!customers.contains(id)) {
+			LOGGER.info("Customer does not exist");
+			return false;
+		}
+		return true;
 	}
 
 }
