+++ src/main/java/com/qa/ims/controller/OrderController.java
@@ -1,5 +1,6 @@
 package com.qa.ims.controller;
 
+import java.sql.Date;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -8,10 +9,14 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+import com.qa.ims.persistence.dao.CustomerDAO;
 import com.qa.ims.persistence.dao.ItemDAO;
 import com.qa.ims.persistence.dao.OrderDAO;
+import com.qa.ims.persistence.dao.OrderItemDAO;
+import com.qa.ims.persistence.domain.Customer;
 import com.qa.ims.persistence.domain.Item;
 import com.qa.ims.persistence.domain.Order;
+import com.qa.ims.persistence.domain.OrderItem;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -53,13 +58,57 @@ public class OrderController implements CrudController<Order> {
     public Order create() {
         LOGGER.info("Please enter the ID of the customer making the order: ");
         Long customerId = utils.getLong();
+
+        if (!checkCustomerID(customerId)) return null;
+
         LOGGER.info("When was the order made? (yyyy-mm-dd): ");
-        LocalDate date = LocalDate.parse(utils.getString());
+        Date date = Date.valueOf(utils.getString());
 
-        Map<Long, Integer> itemQuantities = IdQuantityMap();
-
+        //creating order in order table
         Order order = orderDAO.create(new Order(customerId, date));
+
+        //creating order items in order
+        Map<Long, Integer> itemQuantities = IdQuantityMap();
+        OrderItemDAO orderItemDAO = new OrderItemDAO();
+        for (Map.Entry<Long, Integer> entry : itemQuantities.entrySet()) {
+            Long itemID = entry.getKey();
+            Integer quantity = entry.getValue();
+            OrderItem oi = new OrderItem(order.getID(), itemID, quantity);
+            orderItemDAO.addItem(oi);
+            order.addOrderItem(oi);
+        }
+
         LOGGER.info("Order created.");
+        return order;
+    }
+
+    @Override
+    public Order update() {
+        LOGGER.info("Please enter the ID of the order you'd like to update: ");
+        Long orderID = utils.getLong();
+
+        if (!checkOrderID(orderID)) return null;
+
+        Order order = this.orderDAO.read(orderID);
+
+        OrderItemDAO orderItemDAO = new OrderItemDAO();
+        LOGGER.info("How would you like to update your order?");
+        LOGGER.info("[1] Add an item to the order ");
+        LOGGER.info("[2] Delete an item from the order");
+        LOGGER.info("Enter 1 or 2 to select ");
+        Integer input = utils.getInteger();
+        switch (input) {
+            case 1:
+                addItem(orderID, orderItemDAO, order);
+                return order;
+            case 2:
+                deleteItem(orderID, orderItemDAO, order);
+                break;
+            default:
+                LOGGER.info("Please enter valid input.");
+                break;
+        }
+
         return order;
     }
 
@@ -67,27 +116,34 @@ public class OrderController implements CrudController<Order> {
      * Adds item to an order.
      * @return
      */
-    public Order addItem() {
+    private void addItem(Long orderID, OrderItemDAO orderItemDAO, Order order) {
         LOGGER.info("Please enter the ID of the item you'd like to add: ");
-        Long id = utils.getLong();
-        LOGGER.info("Please enter an item name: ");
-        String name = utils.getString();
-        LOGGER.info("Please enter the value of the item: ");
-        Double value = utils.getDouble();
-        // add to Item List
+        Long itemID = utils.getLong();
+
+        if (!checkItemID(itemID, order)) return;
+
+        LOGGER.info("Item ID: " + itemID + "| How many units of this item are in the order?");
+        Integer quantity = utils.getInteger();
+
+        OrderItem oi = new OrderItem(orderID, itemID, quantity);
+        orderItemDAO.addItem(oi);
+        order.addOrderItem(oi);
         LOGGER.info("Item successfully updated.");
-        return order;
     }
 
     /**
      * Deletes item from an order.
      * @return
      */
-    public Order deleteItem() {
+    private void deleteItem(Long orderID, OrderItemDAO orderItemDAO, Order order) {
         LOGGER.info("Please enter the ID of the item you would like to delete: ");
-        Long id = utils.getLong();
-        // Delete from Item List
-        return orderDAO.delete(id);
+        Long itemID = utils.getLong();
+
+        if (!checkItemID(itemID, order)) return;
+
+        orderItemDAO.deleteItem(orderID, itemID);
+        order.removeOrderItem(orderID, itemID);
+        LOGGER.info("Item successfully deleted.");
     }
 
     /**
@@ -99,6 +155,9 @@ public class OrderController implements CrudController<Order> {
     public int delete() {
         LOGGER.info("Please enter the ID of the order you would like to delete: ");
         Long id = utils.getLong();
+
+        if (!checkOrderID(id)) return 0;
+
         return orderDAO.delete(id);
     }
 
@@ -126,11 +185,38 @@ public class OrderController implements CrudController<Order> {
         List<Integer> quantList = new ArrayList<>();
         for (int i = 0; i < idList.size(); i++) {
             Long current = idList.get(i);
-            LOGGER.info("Item ID: " + current + "| How many of this item are in the order?");
+            LOGGER.info("Item ID: " + current + "| How many units of this item are in the order?");
             quantList.add(utils.getInteger());
         }
 
         return (Map<K, V>) zipToMap(idList, quantList);
+    }
+
+    private boolean checkOrderID(Long id) {
+        List<Order> orderList = orderDAO.readAll();
+        if (!orderList.contains(id)) {
+            LOGGER.info("Order does not exist");
+            return false;
+        }
+        return true;
+    }
+
+    private boolean checkCustomerID(Long id) {
+        CustomerDAO cd = new CustomerDAO();
+        List<Customer> customers = cd.readAll();
+        if (!customers.contains(id)) {
+            LOGGER.info("Customer does not exist");
+            return false;
+        }
+        return true;
+    }
+
+    private boolean checkItemID(Long id, Order order) {
+        if (!order.getItems().contains(id)) {
+            LOGGER.info("Item is not in order");
+            return false;
+        }
+        return true;
     }
 
 }
